Initial division into manager contracts
