+++ contracts/bonding/libraries/Node.sol
@@ -1,0 +1,9 @@
+pragma solidity ^0.4.11;
+
+library Node {
+    struct Node {
+        address id;
+        uint key;
+        bool initialized;
+    }
+}
